Pull out common mock functions, test S3Client
